+++ core/src/main/java/io/cucumber/core/runtime/CucumberExecutionContext.java
@@ -2,7 +2,6 @@ package io.cucumber.core.runtime;
 
 import io.cucumber.core.eventbus.EventBus;
 import io.cucumber.core.exception.CompositeCucumberException;
-import io.cucumber.core.exception.CucumberException;
 import io.cucumber.core.gherkin.Feature;
 import io.cucumber.core.logging.Logger;
 import io.cucumber.core.logging.LoggerFactory;
@@ -23,6 +22,7 @@ import java.util.List;
 import java.util.ResourceBundle;
 import java.util.function.Consumer;
 
+import static io.cucumber.core.exception.ExceptionUtils.printStackTrace;
 import static io.cucumber.core.exception.ExceptionUtils.throwAsUncheckedException;
 import static io.cucumber.createmeta.CreateMeta.createMeta;
 import static io.cucumber.messages.TimeConversion.javaInstantToTimestamp;
@@ -68,36 +68,54 @@ public final class CucumberExecutionContext {
                 .build());
     }
 
+    public void runBeforeAllHooks() {
+        try {
+            runnerSupplier.get().runBeforeAllHooks();
+        } catch (Throwable e) {
+            thrown.add(e);
+            throw e;
+        }
+    }
+
+    public void runAfterAllHooks() {
+        try {
+            runnerSupplier.get().runAfterAllHooks();
+        } catch (Throwable e) {
+            thrown.add(e);
+            throw e;
+        }
+    }
+
     public void finishTestRun() {
         log.debug(() -> "Sending test run finished event");
-        CucumberException cucumberException = getException();
+        Throwable cucumberException = getException();
         emitTestRunFinished(cucumberException);
     }
 
-    public CucumberException getException() {
+    public Throwable getException() {
         if (thrown.isEmpty()) {
             return null;
         }
         if (thrown.size() == 1) {
-            return new CucumberException(thrown.get(0));
+            return thrown.get(0);
         }
         return new CompositeCucumberException(thrown);
     }
 
-    private void emitTestRunFinished(CucumberException cucumberException) {
+    private void emitTestRunFinished(Throwable exception) {
         Instant instant = bus.getInstant();
         Result result = new Result(
-            cucumberException != null ? Status.FAILED : exitStatus.getStatus(),
+            exception != null ? Status.FAILED : exitStatus.getStatus(),
             Duration.between(start, instant),
-            cucumberException);
+            exception);
         bus.send(new TestRunFinished(instant, result));
 
         Messages.TestRunFinished.Builder testRunFinished = Messages.TestRunFinished.newBuilder()
-                .setSuccess(exitStatus.isSuccess())
+                .setSuccess(exception == null && exitStatus.isSuccess())
                 .setTimestamp(javaInstantToTimestamp(instant));
 
-        if (cucumberException != null) {
-            testRunFinished.setMessage(cucumberException.getMessage());
+        if (exception != null) {
+            testRunFinished.setMessage(printStackTrace(exception));
         }
         bus.send(Envelope.newBuilder()
                 .setTestRunFinished(testRunFinished)
@@ -127,7 +145,6 @@ public final class CucumberExecutionContext {
         try {
             return runnerSupplier.get();
         } catch (Throwable e) {
-            log.error(e, () -> "Unable to start Cucumber");
             thrown.add(e);
             throw e;
         }