+++ core/src/main/java/io/cucumber/core/runtime/Runtime.java
@@ -1,7 +1,7 @@
 package io.cucumber.core.runtime;
 
 import io.cucumber.core.eventbus.EventBus;
-import io.cucumber.core.exception.CucumberException;
+import io.cucumber.core.exception.ExceptionUtils;
 import io.cucumber.core.feature.FeatureParser;
 import io.cucumber.core.filter.Filters;
 import io.cucumber.core.gherkin.Feature;
@@ -75,9 +75,27 @@ public final class Runtime {
 
     public void run() {
         context.startTestRun();
-        final List<Feature> features = featureSupplier.get();
+        try {
+            context.runBeforeAllHooks();
+            runFeatures();
+        } finally {
+            try {
+                context.runAfterAllHooks();
+            } finally {
+                context.finishTestRun();
+            }
+        }
+
+        Throwable exception = context.getException();
+        if (exception != null) {
+            ExceptionUtils.throwAsUncheckedException(exception);
+        }
+    }
+
+    private void runFeatures() {
+        List<Feature> features = featureSupplier.get();
         features.forEach(context::beforeFeature);
-        final List<Future<?>> executingPickles = features.stream()
+        List<Future<?>> executingPickles = features.stream()
                 .flatMap(feature -> feature.getPickles().stream())
                 .filter(filter)
                 .collect(collectingAndThen(toList(),
@@ -97,12 +115,6 @@ public final class Runtime {
                 executor.shutdownNow();
                 log.debug(e, () -> "Interrupted while executing pickle");
             }
-        }
-        context.finishTestRun();
-
-        CucumberException exception = context.getException();
-        if (exception != null) {
-            throw exception;
         }
     }