+++ junit/src/main/java/io/cucumber/junit/Cucumber.java
@@ -92,7 +92,6 @@ public final class Cucumber extends ParentRunner<ParentRunner<?>> {
 
     private final List<ParentRunner<?>> children;
     private final EventBus bus;
-    private final List<Feature> features;
     private final Plugins plugins;
     private final CucumberExecutionContext context;
 
@@ -155,7 +154,7 @@ public final class Cucumber extends ParentRunner<ParentRunner<?>> {
         Supplier<ClassLoader> classLoader = ClassLoaders::getDefaultClassLoader;
         FeaturePathFeatureSupplier featureSupplier = new FeaturePathFeatureSupplier(classLoader, runtimeOptions,
             parser);
-        this.features = featureSupplier.get();
+        List<Feature> features = featureSupplier.get();
 
         // Create plugins after feature parsing to avoid the creation of empty
         // files on lexer errors.
@@ -179,7 +178,7 @@ public final class Cucumber extends ParentRunner<ParentRunner<?>> {
         this.children = features.stream()
                 .map(feature -> {
                     Integer uniqueSuffix = uniqueSuffix(groupedByName, feature, Feature::getName);
-                    return FeatureRunner.create(feature, uniqueSuffix, filters, runnerSupplier, junitOptions);
+                    return FeatureRunner.create(feature, uniqueSuffix, filters, context, junitOptions);
                 })
                 .filter(runner -> !runner.isEmpty())
                 .collect(toList());
@@ -202,8 +201,15 @@ public final class Cucumber extends ParentRunner<ParentRunner<?>> {
 
     @Override
     protected Statement childrenInvoker(RunNotifier notifier) {
-        Statement runFeatures = super.childrenInvoker(notifier);
-        return new RunCucumber(runFeatures);
+        Statement statement = super.childrenInvoker(notifier);
+
+        statement = new RunBeforeAllHooks(statement);
+        statement = new RunAfterAllHooks(statement);
+
+        statement = new StartTestRun(statement);
+        statement = new FinishTestRun(statement);
+
+        return statement;
     }
 
     @Override
@@ -212,12 +218,11 @@ public final class Cucumber extends ParentRunner<ParentRunner<?>> {
         multiThreadingAssumed = true;
     }
 
-    class RunCucumber extends Statement {
+    private class StartTestRun extends Statement {
+        private final Statement next;
 
-        private final Statement runFeatures;
-
-        RunCucumber(Statement runFeatures) {
-            this.runFeatures = runFeatures;
+        public StartTestRun(Statement next) {
+            this.next = next;
         }
 
         @Override
@@ -227,14 +232,58 @@ public final class Cucumber extends ParentRunner<ParentRunner<?>> {
             } else {
                 plugins.setEventBusOnEventListenerPlugins(bus);
             }
-
             context.startTestRun();
-            features.forEach(context::beforeFeature);
+            next.evaluate();
+        }
 
+    }
+
+    private class FinishTestRun extends Statement {
+        private final Statement next;
+
+        public FinishTestRun(Statement next) {
+            this.next = next;
+        }
+
+        @Override
+        public void evaluate() throws Throwable {
             try {
-                runFeatures.evaluate();
+                next.evaluate();
             } finally {
                 context.finishTestRun();
+            }
+        }
+
+    }
+
+    private class RunBeforeAllHooks extends Statement {
+        private final Statement next;
+
+        public RunBeforeAllHooks(Statement next) {
+            this.next = next;
+        }
+
+        @Override
+        public void evaluate() throws Throwable {
+            context.runBeforeAllHooks();
+            next.evaluate();
+        }
+
+    }
+
+    private class RunAfterAllHooks extends Statement {
+        private final Statement next;
+
+        public RunAfterAllHooks(Statement next) {
+            this.next = next;
+        }
+
+        @Override
+        public void evaluate() throws Throwable {
+            try {
+                next.evaluate();
+            } finally {
+                context.runAfterAllHooks();
             }
         }